// packages/partridge-client-core/src/__types__/graphql/__generated__/types.ts
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any
  /**
   * A signed eight-byte integer. The upper big integer values are greater than the
   * max value for a JavaScript number. Therefore all big integers will be output as
   * strings and not numbers.
   **/
  BigInt: any
  /** A universally unique identifier as defined by [RFC 4122](https://tools.ietf.org/html/rfc4122). */
  UUID: any
  /**
   * A point in time as described by the [ISO
   * 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone.
   **/
  Datetime: any
  /** A JavaScript object encoded in the JSON format as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any
}
// ...
export type DimensionsInput = {
  width: Scalars['Int']
  height: Scalars['Int']
  scale: Scalars['Int']
  fontScale: Scalars['Int']
  lessThan: DimensionsSetInput
  greaterThan: DimensionsSetInput
  is: DimensionsSetInput
  orientation: Orientation
}

export type DimensionsSet = {
  __typename?: 'DimensionsSet'
  xs: Scalars['Boolean']
  sm: Scalars['Boolean']
  md: Scalars['Boolean']
  lg: Scalars['Boolean']
  xl: Scalars['Boolean']
  xxl: Scalars['Boolean']
}

export type DimensionsSetInput = {
  xs: Scalars['Boolean']
  sm: Scalars['Boolean']
  md: Scalars['Boolean']
  lg: Scalars['Boolean']
  xl: Scalars['Boolean']
  xxl: Scalars['Boolean']
}
// ...
export enum Orientation {
  Portrait = 'PORTRAIT',
  Landscape = 'LANDSCAPE',
}